Replaced jshint with eslint and fixed all warning/errors
eslint is a lot more comprehensive and shows failed conditions

+++ controllers/game-controller.js
@@ -25,18 +25,18 @@ class GameController {
         this.colorService = new ColorService();
         this.nameService = new NameService();
         this.playerStatBoard = new PlayerStatBoard();
-        this.adminService = new AdminService(this.players,this.food, this.playerStatBoard, 
+        this.adminService = new AdminService(this.players, this.food, this.playerStatBoard,
             this.boardOccupancyService, this.colorService, this.nameService, this.playerSpawnService,
             this.generateFood.bind(this), this.removeFood.bind(this), this._disconnectPlayer.bind(this),
             this.sendNotificationToPlayers.bind(this));
-        
-        for(let i = 0; i < ServerConfig.DEFAULT_FOOD_AMOUNT; i++) {
+
+        for (let i = 0; i < ServerConfig.DEFAULT_FOOD_AMOUNT; i++) {
             this.generateFood();
         }
-        
+
         this.io = io;
         const self = this;
-        this.io.sockets.on(ServerConfig.IO.DEFAULT_CONNECTION, function (socket) {
+        this.io.sockets.on(ServerConfig.IO.DEFAULT_CONNECTION, socket => {
             socket.on(ServerConfig.IO.INCOMING.NEW_PLAYER, self._addPlayer.bind(self, socket));
             socket.on(ServerConfig.IO.INCOMING.NAME_CHANGE, self._changePlayerName.bind(self, socket));
             socket.on(ServerConfig.IO.INCOMING.COLOR_CHANGE, self._changeColor.bind(self, socket));
@@ -48,57 +48,59 @@ class GameController {
             socket.on(ServerConfig.IO.INCOMING.JOIN_GAME, self._playerJoinGame.bind(self, socket));
             socket.on(ServerConfig.IO.INCOMING.SPECTATE_GAME, self._playerSpectateGame.bind(self, socket));
             socket.on(ServerConfig.IO.INCOMING.DISCONNECT, self._disconnect.bind(self, socket));
-            
-            socket.on(ServerConfig.IO.INCOMING.BOT_CHANGE, 
+
+            socket.on(ServerConfig.IO.INCOMING.BOT_CHANGE,
                 self.adminService.changeBots.bind(self.adminService, socket));
-            socket.on(ServerConfig.IO.INCOMING.FOOD_CHANGE, 
+            socket.on(ServerConfig.IO.INCOMING.FOOD_CHANGE,
                 self.adminService.changeFood.bind(self.adminService, socket));
-            socket.on(ServerConfig.IO.INCOMING.SPEED_CHANGE, 
+            socket.on(ServerConfig.IO.INCOMING.SPEED_CHANGE,
                 self.adminService.changeSpeed.bind(self.adminService, socket));
-            socket.on(ServerConfig.IO.INCOMING.START_LENGTH_CHANGE, 
+            socket.on(ServerConfig.IO.INCOMING.START_LENGTH_CHANGE,
                 self.adminService.changeStartLength.bind(self.adminService, socket));
         });
     }
-    
+
     runGameCycle() {
         // Pause and reset the game if there aren't any players
-        if(Object.keys(this.players).length - this.adminService.getBotNames().length === 0){
+        if (Object.keys(this.players).length - this.adminService.getBotNames().length === 0) {
             console.log("Game Paused");
             this.adminService.resetGame();
             return;
         }
-        
-        for(const botName of this.adminService.getBotNames()) {
+
+        for (const botName of this.adminService.getBotNames()) {
             const bot = this.players[botName];
-            if(Math.random() <= ServerConfig.BOT_CHANGE_DIRECTION_PERCENT) {
+            if (Math.random() <= ServerConfig.BOT_CHANGE_DIRECTION_PERCENT) {
                 this.botDirectionService.changeToRandomDirection(bot);
             }
             this.botDirectionService.changeDirectionIfInDanger(bot);
         }
-        
+
         const playersToRespawn = [];
-        for(const playerId in this.players) {
-            const player = this.players[playerId];
-            // Check if player is spectating
-            if(player.segments.length === 0) {
-                continue;
-            }
-            this.boardOccupancyService.removePlayerOccupancy(player.id, player.segments);
-            CoordinateService.movePlayer(player);
-            if(this.boardOccupancyService.isOutOfBounds(player.getHeadLocation()) ||
-                    this.boardOccupancyService.isWall(player.getHeadLocation())) {
-                player.clearAllSegments();
-                playersToRespawn.push(player);
-            } else {
-                this.boardOccupancyService.addPlayerOccupancy(player.id, player.segments);
+        for (const playerId in this.players) {
+            if ({}.hasOwnProperty.call(this.players, playerId)) {
+                const player = this.players[playerId];
+                // Check if player is spectating
+                if (player.segments.length === 0) {
+                    continue;
+                }
+                this.boardOccupancyService.removePlayerOccupancy(player.id, player.segments);
+                CoordinateService.movePlayer(player);
+                if (this.boardOccupancyService.isOutOfBounds(player.getHeadLocation()) ||
+                        this.boardOccupancyService.isWall(player.getHeadLocation())) {
+                    player.clearAllSegments();
+                    playersToRespawn.push(player);
+                } else {
+                    this.boardOccupancyService.addPlayerOccupancy(player.id, player.segments);
+                }
             }
         }
-        
+
         const killReports = this.boardOccupancyService.getKillReports();
-        for(const killReport of killReports) {
-            if(killReport.isSingleKill()) {
+        for (const killReport of killReports) {
+            if (killReport.isSingleKill()) {
                 const victim = this.players[killReport.victimId];
-                if(killReport.killerId === killReport.victimId) {
+                if (killReport.killerId === killReport.victimId) {
                     // TODO Display suicide announcement
                 } else {
                     this.playerStatBoard.addKill(killReport.killerId);
@@ -110,51 +112,51 @@ class GameController {
                 victim.clearAllSegments();
                 playersToRespawn.push(victim);
             } else {
-                for(const victimId of killReport.victimIds) {
+                for (const victimId of killReport.victimIds) {
                     const victim = this.players[victimId];
                     this.boardOccupancyService.removePlayerOccupancy(victim.id, victim.segments);
                     victim.clearAllSegments();
-                    playersToRespawn.push(victim); 
+                    playersToRespawn.push(victim);
                 }
                 // TODO Display multideath announcement
             }
         }
-        
-        for(const player of playersToRespawn) {
+
+        for (const player of playersToRespawn) {
             this.respawnPlayer(player);
         }
-        
+
         let foodToRespawn = 0;
         const foodsConsumed = this.boardOccupancyService.getFoodsConsumed();
-        for(const foodConsumed of foodsConsumed) {
+        for (const foodConsumed of foodsConsumed) {
             const playerWhoConsumedFood = this.players[foodConsumed.playerId];
             this.removeFood(foodConsumed.foodId);
-            
+
             playerWhoConsumedFood.growNextTurn();
             this.playerStatBoard.increaseScore(playerWhoConsumedFood.id);
             foodToRespawn++;
         }
-        
-        for(let i = 0; i < foodToRespawn; i++) {
+
+        for (let i = 0; i < foodToRespawn; i++) {
             this.generateFood();
         }
-    
+
         const gameData = {
             players: this.players,
             food: this.food,
             playerStats: this.playerStatBoard,
             speed: this.adminService.getGameSpeed(),
             numberOfBots: this.adminService.getBotNames().length,
-            startLength: this.adminService.getPlayerStartLength()
+            startLength: this.adminService.getPlayerStartLength(),
         };
-        this.io.sockets.emit(ServerConfig.IO.OUTGOING.NEW_STATE, gameData );
-        
-        setTimeout(this.runGameCycle.bind(this), 1000/this.adminService.getGameSpeed());
+        this.io.sockets.emit(ServerConfig.IO.OUTGOING.NEW_STATE, gameData);
+
+        setTimeout(this.runGameCycle.bind(this), 1000 / this.adminService.getGameSpeed());
     }
-    
+
     generateFood() {
         const randomUnoccupiedCoordinate = this.boardOccupancyService.getRandomUnoccupiedCoordinate();
-        if(!randomUnoccupiedCoordinate) {
+        if (!randomUnoccupiedCoordinate) {
             this.sendNotificationToPlayers("Could not add more food.  No room left.", "white");
             return;
         }
@@ -163,30 +165,30 @@ class GameController {
         this.food[foodId] = food;
         this.boardOccupancyService.addFoodOccupancy(food.id, food.location);
     }
-    
+
     removeFood(foodId) {
         const foodToRemove = this.food[foodId];
         this.nameService.returnFoodId(foodId);
         this.boardOccupancyService.removeFoodOccupancy(foodId, foodToRemove.location);
         delete this.food[foodId];
     }
-    
+
     respawnPlayer(player) {
         this.playerSpawnService.setupNewSpawn(player, this.adminService.getPlayerStartLength(),
             ServerConfig.SPAWN_TURN_LEEWAY);
         this.playerStatBoard.resetScore(player.id);
         this.playerStatBoard.addDeath(player.id);
     }
-    
+
     sendNotificationToPlayers(notification, playerColor) {
         console.log(notification);
-        this.io.sockets.emit(ServerConfig.IO.OUTGOING.NOTIFICATION, notification, playerColor );
+        this.io.sockets.emit(ServerConfig.IO.OUTGOING.NOTIFICATION, notification, playerColor);
     }
-    
+
     /*******************************
      *  socket.io handling methods *
      *******************************/
-    
+
     _addPlayer(socket, previousName, previousImage) {
         const playerName = this.nameService.getPlayerName();
         const playerColor = this.colorService.getColor();
@@ -197,22 +199,22 @@ class GameController {
         this.playerStatBoard.addPlayer(newPlayer.id, playerName, playerColor);
         socket.emit(ServerConfig.IO.OUTGOING.NEW_PLAYER_INFO, playerName, playerColor);
         socket.emit(ServerConfig.IO.OUTGOING.BOARD_INFO, Board);
-        this.sendNotificationToPlayers(playerName + " has joined!", playerColor);
-        
-        if(previousName) {
+        this.sendNotificationToPlayers(`${playerName} has joined!`, playerColor);
+
+        if (previousName) {
             this._changePlayerName(socket, previousName);
         }
-        if(previousImage) {
+        if (previousImage) {
             this._updatePlayerImage(socket, previousImage);
         }
-        
+
         // Start game if the first player has joined
-        if(Object.keys(this.players).length === 1) {
+        if (Object.keys(this.players).length === 1) {
             console.log("Game Started");
             this.runGameCycle();
         }
     }
-    
+
     _changeColor(socket) {
         const player = this.players[socket.id];
         const newColor = this.colorService.getColor();
@@ -220,88 +222,88 @@ class GameController {
         player.color = newColor;
         this.playerStatBoard.changePlayerColor(player.id, newColor);
         socket.emit(ServerConfig.IO.OUTGOING.NEW_PLAYER_INFO, player.name, newColor);
-        this.sendNotificationToPlayers(player.name + " has changed colors.", newColor);
+        this.sendNotificationToPlayers(`${player.name} has changed colors.`, newColor);
     }
-    
+
     _changePlayerName(socket, newPlayerName) {
         const player = this.players[socket.id];
         const oldPlayerName = player.name;
-        newPlayerName = newPlayerName.trim();
-        if(oldPlayerName === newPlayerName) {
+        const trimmedNewPlayerName = newPlayerName.trim();
+        if (oldPlayerName === trimmedNewPlayerName) {
             return;
         }
-        if(this.nameService.doesPlayerNameExist(newPlayerName)) {
+        if (this.nameService.doesPlayerNameExist(trimmedNewPlayerName)) {
             socket.emit(ServerConfig.IO.OUTGOING.NEW_PLAYER_INFO, oldPlayerName, player.color);
-            this.sendNotificationToPlayers(player.name + " couldn't claim the name " + newPlayerName, player.color);
+            this.sendNotificationToPlayers(`${player.name} couldn't claim the name ${trimmedNewPlayerName}`, player.color);
         } else {
-            this.sendNotificationToPlayers(oldPlayerName + " is now known as " + newPlayerName, player.color);
-            player.name = newPlayerName;
-            this.nameService.usePlayerName(newPlayerName);
-            this.playerStatBoard.changePlayerName(player.id, newPlayerName);
-            socket.emit(ServerConfig.IO.OUTGOING.NEW_PLAYER_INFO, newPlayerName, player.color);
+            this.sendNotificationToPlayers(`${oldPlayerName} is now known as ${trimmedNewPlayerName}`, player.color);
+            player.name = trimmedNewPlayerName;
+            this.nameService.usePlayerName(trimmedNewPlayerName);
+            this.playerStatBoard.changePlayerName(player.id, trimmedNewPlayerName);
+            socket.emit(ServerConfig.IO.OUTGOING.NEW_PLAYER_INFO, trimmedNewPlayerName, player.color);
         }
     }
-    
+
     _disconnect(socket) {
         this._disconnectPlayer(socket.id);
     }
-    
+
     _disconnectPlayer(playerId) {
         const player = this.players[playerId];
-        if(!player) {
+        if (!player) {
             return;
         }
-        this.sendNotificationToPlayers(player.name + " has left.", player.color);
+        this.sendNotificationToPlayers(`${player.name} has left.`, player.color);
         this.colorService.returnColor(player.color);
         this.nameService.returnPlayerName(player.name);
         this.playerStatBoard.removePlayer(player.id);
-        if(player.segments.length > 0) {
+        if (player.segments.length > 0) {
             this.boardOccupancyService.removePlayerOccupancy(player.id, player.segments);
         }
         delete this.players[playerId];
     }
-    
+
     _keyDown(socket, keyCode) {
         GameControlsService.handleKeyDown(this.players[socket.id], keyCode);
     }
-    
+
     _playerJoinGame(socket) {
         const player = this.players[socket.id];
         this.respawnPlayer(player);
-        this.sendNotificationToPlayers(player.name + " has rejoined the game.", player.color);
+        this.sendNotificationToPlayers(`${player.name} has rejoined the game.`, player.color);
     }
-    
+
     _playerSpectateGame(socket) {
         const player = this.players[socket.id];
         this.boardOccupancyService.removePlayerOccupancy(player.id, player.segments);
         player.clearAllSegments();
-        this.sendNotificationToPlayers(player.name + " is now spectating.", player.color);
+        this.sendNotificationToPlayers(`${player.name} is now spectating.`, player.color);
     }
-    
+
     _clearBackgroundImage(socket) {
         const player = this.players[socket.id];
-        this.io.sockets.emit(ServerConfig.IO.OUTGOING.NEW_BACKGROUND_IMAGE );
-        this.sendNotificationToPlayers(player.name + " has clear the background image.", player.color);
+        this.io.sockets.emit(ServerConfig.IO.OUTGOING.NEW_BACKGROUND_IMAGE);
+        this.sendNotificationToPlayers(`${player.name} has clear the background image.`, player.color);
     }
-    
+
     _clearPlayerImage(socket) {
         const player = this.players[socket.id];
         delete player.base64Image;
         this.playerStatBoard.clearPlayerImage(player.id);
-        this.sendNotificationToPlayers(player.name + " has removed their image.", player.color);
+        this.sendNotificationToPlayers(`${player.name} has removed their image.`, player.color);
     }
-    
+
     _updateBackgroundImage(socket, base64Image) {
         const player = this.players[socket.id];
-        this.io.sockets.emit(ServerConfig.IO.OUTGOING.NEW_BACKGROUND_IMAGE, base64Image );
-        this.sendNotificationToPlayers(player.name + " has updated the background image.", player.color);
+        this.io.sockets.emit(ServerConfig.IO.OUTGOING.NEW_BACKGROUND_IMAGE, base64Image);
+        this.sendNotificationToPlayers(`${player.name} has updated the background image.`, player.color);
     }
-    
+
     _updatePlayerImage(socket, base64Image) {
         const player = this.players[socket.id];
         player.setBase64Image(base64Image);
         this.playerStatBoard.setBase64Image(player.id, base64Image);
-        this.sendNotificationToPlayers(player.name + " has uploaded a new image.", player.color);
+        this.sendNotificationToPlayers(`${player.name} has uploaded a new image.`, player.color);
     }
 }
